Make sure original_data is a dict
Item._original_data is not necessarily a dict -- for example, Config
stores a list of dicts in it.

Linting parameters requires the data to be a dict, so add a check for
the content type. Ignoring non-dicts should not change any functionality,
as the whole config isn't passed to parameter parsing, but it allows
adding better type hinting while also making the type checker happy.