Improve config parser type hints
Config.parse does not work unless the data is an iterable container of
dicts. Add a type hint to show this, and improve the error message for a
list item that is not a dict.

Add a generic type hint form Item._original_data to make mypy happier
about the whole thing.