Sort selected test snapshot lines
Sometimes, test_validation failed on some environments due to the output
lines being in a different order than the snapshot.

Since the content of the validation error is what is important, it
should be enough to just test that both contents have the same lines, no
matter which order the messages are in.

The sorting could maybe have been made mandatory, but since the
flakiness has only appeared in one of the tests, make it optional to fix
just that test and leave the rest be. This can be changed if flakiness
is found in other tests later.