Add linter warning about duplicate names
Sometimes you can accidentally use the same name for two steps, tasks,
etc. -- especially when copying and pasting configurations. This will
work, but only the latter definition will be used.

As this is probably almost always not the wanted effect, make the linter
warn about the duplicates (but not make it an error, as technically it
is allowed).

# Conflicts:
#	tests/__snapshots__/test_validation.ambr